CPS-201: Add separate class for Workflow Menu items
